+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.7.0-beta.0] - Unreleased
 
+### Added
+- Added Erlang distribution over serial (uart)
+
 ## [0.7.0-alpha.1] - 2026-04-06
 
 ### Added